+++ RevitLookup.UI/Appearance/Theme.cs
@@ -35,15 +35,16 @@ public static class Theme
     /// <summary>
     /// Changes the current application theme.
     /// </summary>
+    /// <param name="element">Framework element</param>
     /// <param name="themeType">Theme to set.</param>
     /// <param name="backgroundEffect">Whether the custom background effect should be applied.</param>
     /// <param name="updateAccent">Whether the color accents should be changed.</param>
     /// <param name="forceBackground">If <see langword="true"/>, bypasses the app's theme compatibility check and tries to force the change of a background effect.</param>
-    public static void Apply(ThemeType themeType, WindowBackdropType backgroundEffect = WindowBackdropType.Mica,
+    public static void Apply(FrameworkElement element, ThemeType themeType, WindowBackdropType backgroundEffect = WindowBackdropType.Mica,
         bool updateAccent = true, bool forceBackground = false)
     {
         if (updateAccent)
-            Accent.Apply(
+            Accent.Apply(element,
                 Accent.GetColorizationColor(),
                 themeType,
                 false