+++ RevitLookup/Views/Pages/EventsView.xaml.cs
@@ -22,6 +22,7 @@ using System.Windows.Data;
 using RevitLookup.Core.Objects;
 using RevitLookup.Services.Contracts;
 using RevitLookup.ViewModels.Pages;
+using Wpf.Ui.Appearance;
 
 namespace RevitLookup.Views.Pages;
 
@@ -41,5 +42,6 @@ public sealed partial class EventsView
 
         ViewModel.SearchResultsChanged += OnSearchResultsChanged;
         TreeView.SelectedItemChanged += OnTreeSelectionChanged;
+        Theme.Apply(this, settingsService.Theme, settingsService.Background);
     }
 }